+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Adds copyrightsByLicense field to the attribution report JSON output.
 - Removes copyright information from attribution reports in the JSON output ([#945](https://github.com/fossas/fossa-cli/pull/945)) as it was never available from the server.
 
 ## v3.3.3